+++ crates/wdl-gauntlet/src/lib.rs
@@ -40,8 +40,7 @@ use report::UnmatchedStatus;
 pub use repository::Repository;
 use wdl_lint::ast::Document;
 use wdl_lint::ast::Validator;
-use wdl_lint::rules;
-use wdl_lint::ExceptVisitor;
+use wdl_lint::LintVisitor;
 
 use crate::repository::WorkDir;
 
@@ -185,8 +184,7 @@ pub async fn gauntlet(args: Args) -> Result<()> {
                 Ok(document) => {
                     let mut validator = Validator::default();
                     if args.arena {
-                        validator
-                            .add_visitor(ExceptVisitor::new(rules().iter().map(AsRef::as_ref)));
+                        validator.add_visitor(LintVisitor::default());
                     }
 
                     match validator.validate(&document) {