+++ crates/wdl-lint/tests/lints.rs
@@ -31,8 +31,7 @@ use rayon::prelude::*;
 use wdl_ast::Diagnostic;
 use wdl_ast::Document;
 use wdl_ast::Validator;
-use wdl_lint::rules;
-use wdl_lint::ExceptVisitor;
+use wdl_lint::LintVisitor;
 
 fn find_tests() -> Vec<PathBuf> {
     // Check for filter arguments consisting of test names
@@ -130,7 +129,7 @@ fn run_test(test: &Path, ntests: &AtomicUsize) -> Result<(), String> {
     match Document::parse(&source).into_result() {
         Ok(document) => {
             let mut validator = Validator::default();
-            validator.add_visitor(ExceptVisitor::new(rules().iter().map(AsRef::as_ref)));
+            validator.add_visitor(LintVisitor::default());
             let errors = match validator.validate(&document) {
                 Ok(()) => String::new(),
                 Err(diagnostics) => format_diagnostics(&diagnostics, &path, &source),