+++ crates/wdl-lint/CHANGELOG.md
@@ -28,8 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-### Changed
-
+* Refactored the lint rules so that they directly implement `Visitor`; renamed
+  `ExceptVisitor` to `LintVisitor` ([#103](https://github.com/stjude-rust-labs/wdl/pull/103)).
 * Refactored the lint rules so that they are not in a `v1` module
   ([#95](https://github.com/stjude-rust-labs/wdl/pull/95)).